+++ CHANGELOG.md
@@ -1,15 +1,17 @@
 # Changelog
 
-## 6.3.0 (unreleased)
+## 4.3.0 (unreleased)
 * **Features**
     * add git describe version placeholders
-      enable describe feature by set config `describeTagPattern` at root level or at branch/tag config level
         * new placeholders
             * `${describe}`
             * `${describe.tag}`
-                * `${describe.<TAG_PATTERN_GROUP_NAME or TAG_PATTERN_GROUP_INDEX>}`
+                * `${describe.<TAG_PATTERN_GROUP_NAME or TAG_PATTERN_GROUP_INDEX>}` e.g. pattern `v(?<version>.*)` will create placeholder `${describe.version}`
             * `${describe.distance}`
 
+* **BREAKING CHANGES**
+    * no longer provide project property `git.dirty` due to performance issues on larger projects,
+      version format placeholder `${dirty}` is still available
     
 ## 4.1.0
 * **Features**